Use ConcurrentDictionary
Dictionary is not thread safe.
With multiple threads we can wind up in the situation where:

Thread A: Check to see if item is in dictionary -> false.
Thread B: Check to see if item is in dictionary -> false,
Thread A: Add item to dictionary -> success,
Thread B: Add item to dictionary -> Item Already Exists, Throw Exception

ConcurrentDictionary provides us with TryAdd which will attempt to add
the item to the dictionary and will not throw if the item already
exists.